Add DRM backend display driver
This adds a display driver using the libdrm to target DRM/KMS systems.

It handles double buffering over Atomic commits, and does a modeset for
the first commit.

It can handle the page flip event in the wait_cb callback.

The lv_drv_conf has an optional setting to select the connector if the
system has multiple connectors.

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>