feat(libinput/evdev) add full keyboard support (#156)
This enhances the existing libinput and evdev driver by adding support
for full keyboards. The conversion from Linux scan codes to UTF-8
characters is performed via XKB. To control the behavior a new config
setting for opting into or out of using XKB was added. The keymap can
either be set via another config option or by calling a function at
runtime.

Two small caveats:

(1) This only works for a single connected keyboard because static
variables are used to store the keyboard state. Supporting multiple
devices of the same type is still being discussed in #151.
(2) When entering non-ASCII characters, the selected font needs to
include the respective glyph, otherwise it won't render on the UI.

Fixes: #153