Add support for devices with KEY_POWER
This makes it possible to automatically locate power buttons (which have
KEY_POWER) via libinput. Additionally, a new LV_KEY_POWER key is
introduced that is dispatched whenever KEY_POWER is pressed.

In order to handle LV_KEY_POWER, the client needs to register an indev
with type LV_INDEV_TYPE_KEYPAD and set the feedback_cb callback.
Additionally, the indev needs to be assigned a group with at least one
element.

Fixes: #161