Make it possible to use multiple devices with the libinput and XKB drivers
This change bundles all of the previously file-global device-specific
state in libinput.c and xkb.c into new driver state structs with a
default value. All of the existing public functions are retained and
operate on the default struct values. To allow the use of multiple
devices, a second set of public functions is introduced that also take
the specific driver state as an argument. It is left up to the caller
where to store the driver state and how to retrieve it when reading
input events.

Closes: #151