Use LV_LOG instead of printf in fbdev driver
This replaces the previous `printf` calls with `LV_LOG_INFO`. When using
LVGL in an application that has to forward STDOUT to another process,
this is required so that logging output from the fbdev driver can be
controlled with the LV_LOG settings to prevent feeding it to other
programs.

See also: https://gitlab.com/cherrypicker/unl0kr/-/issues/16