libinput: insert dummy events to handle fast OSK typing
When typing fast with thumbs, you might hit a key while your other thumb
is already pressing a key, however this confuses indev quite a lot.

For example, you'll press a key (say 'h') with your right thumb, then
before releasing your thumb you'll press the 'e' key with your left
thumb, and then finally release the 'h' key followed by the 'e' key.

This should result in "he" being typed, but usually results in an "e".
Indev assumes that a pointer release will always be for whatever the
last position was. This is reasonable for mice but not for multitouch
displays.

This workaround detects this scenario and inserts two dummy events,
first a "pressed" state with the coordinates of the first finger (the
'h' key), then a release event, triggering the keypress, then it sends
another pressed events with the position of the second finger (the 'e'
key) so that if the next event is that finger releasing, the position
will be correct.