fix(build): Update xpt2046_read to match signature required by lv_indev_drv_t's read callback. (#290)
The indev read callback previously returned bool to indicate whether it should
be called again. Per the changelog, this was modified to return this indicator
in the lv_indev_data_t struct's continue_reading member. This change updates
the xpt2046_read function to be compatible with this new calling convention.

BREAKING CHANGE

Co-authored-by: John Auld <john@johnauld.com>